+++ JustSaying/JustSayingFluentlyExtensions.cs
@@ -1,0 +1,10 @@
+namespace JustSaying
+{
+    public static class JustSayingFluentlyExtensions
+    {
+        public static IFluentSubscription IntoDefaultQueue(this ISubscriberIntoQueue subscriber)
+        {
+            return subscriber.IntoQueue(string.Empty);
+        }
+    }
+}
